% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.arg.MB.R
\name{check.arg.MB}
\alias{check.arg.MB}
\title{check.arg.MB}
\usage{
check.arg.MB(data, nature, rank.restriction, blocs, blocs.name)
}
\arguments{
\item{data}{raw variables (vector or matrix)}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{rank.restriction}{restriction of the quantification matrix for nominal variable}

\item{blocs}{vector(length k) with number of variables in each bloc}

\item{blocs.name}{vector(length k) with names of each bloc}
}
\value{
stop function if argument or data ar wrong
}
\description{
check if data and arguments are good to be analysed by MBPCAOS function
}
