% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.arg.R
\name{check.arg}
\alias{check.arg}
\title{check.arg}
\usage{
check.arg(data, nature, rank.restriction)
}
\arguments{
\item{data}{raw variables (vector or matrix)}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{rank.restriction}{restriction of the quantification matrix for nominal variable}
}
\value{
stop function if argument or data ar wrong
}
\description{
check if data and arguments are good to be analysed by PCAOS function
}
