% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.plot.MB.arg.R
\name{check.plot.MB.arg}
\alias{check.plot.MB.arg}
\title{check.plot.MB.arg}
\usage{
check.plot.MB.arg(choice, nature, nature.supp, supp.var, comp, nb.comp)
}
\arguments{
\item{choice}{the graph to plot possible values are "screeplot","quantif","indiv","cor","modalities","mixed","squared loadings". See Details.}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{nature.supp}{nature of supplementary variables}

\item{supp.var}{if the user asked to plot the supplementary variable or not}

\item{comp}{axes to use in the plots}

\item{nb.comp}{number of components in the PCAOS model}
}
\value{
stop function if argument are wrong
}
\description{
check if data and arguments are good to be plot for (MBPCAOS)
}
