% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.component.R
\name{choice.component}
\alias{choice.component}
\title{choice.component}
\usage{
choice.component(
  data,
  nature = rep("num", ncol(data)),
  nb.comp.to.investigate = ncol(data)
)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{nb.comp.to.investigate}{Number of components to investigate (default=ncol(data))}
}
\value{
Data frame with global Loss value and percentage of the quantified variables inertia explained, for different model dimensionality (H). The last column provides the variation in percentage between to two successive size of dimensionality.
}
\description{
Helps user choose appropriate number of component to optimise for PCAOS method.
}
