% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.coding.R
\name{dummy.coding}
\alias{dummy.coding}
\title{dummy.coding}
\usage{
dummy.coding(data)
}
\arguments{
\item{data}{raw variables (vector or matrix)}
}
\value{
dummy coded variables
}
\description{
dummy.coding
}
