% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.weight.R
\name{freq.weight}
\alias{freq.weight}
\title{freq.weight}
\usage{
freq.weight(dummy)
}
\arguments{
\item{dummy}{dummy coded variables}
}
\value{
pretreated dummy coded variables by khi-deux metrics
}
\description{
freq.weight
}
