% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBPCAOS.R
\name{MBPCAOS}
\alias{MBPCAOS}
\title{MultiBlock Principal Components Analysis with Optimal Scaling features}
\usage{
MBPCAOS(
  data,
  level.scale = rep("num", ncol(data)),
  blocs,
  blocs.name = paste("Bloc", 1:length(blocs)),
  block.scaling = "inertia",
  nb.comp = 2,
  maxiter = 100,
  threshold = 1e-06,
  supp.var = NULL,
  print.order = TRUE,
  init = "rdm"
)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{level.scale}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{blocs}{vector(length k) with number of variables in each bloc}

\item{blocs.name}{vector(length k) with names of each bloc}

\item{block.scaling}{scaling applied to each block. Possible value are : \itemize{
  \item "inertia"(default): each quantified block is divided by its total inertia (sum of square).
  \item "lambda1" : each quantified block is divided by its the first singular value.
  \item "null" : no scaling is applied
}}

\item{nb.comp}{number of components of the model (by default 2)}

\item{maxiter}{maximum number of iterations.}

\item{threshold}{the threshold for assessing convergence}

\item{supp.var}{a vector indicating the indexes of the supplementary variables}

\item{print.order}{boolean (TRUE by default), if TRUE ther order of the categories of ordinal variables are print}

\item{init}{Intitialization strategy, possible values are :
\itemize{
  \item "rdm" (default) : random initialisation
  \item "svd": components are initialized with the singular value decomposition of the concatenated and pretreated variables (numeric variables are standardized and categorical variables are coded as pseudo disjunctive and weighted by the frequencies of the categories)
}}
}
\value{
\itemize{
  \item weigths : list of weights of the variables (loadings and weights are the same in PCA like model)
  \item components : data.frame with individuals scores on each dimension.
  \item quantified.data : Optimally quantified variables
  \item quantified.data.list : Optimally quantified blocks (list form)
  \item block.components : components associated with each block
  \item summary : summary of number of variables and it's nature
  \item quantification.categories.nom : list of optimally quantified categories (nominal variables)
  \item quantification.categories.ord : list of optimally quantified categories (ordinal variables)
  \item inertia : percentage and cumulative percentage of variance of the quantified variables explained
  \item stockiter : evolution of criterion for each ieration
  \item data : orginal dataset
  \item level.scale : level.scale of scaling choosen for each variable
  \item blocs : number of variable in each block
  \item blocs.name : name of each block
}
}
\description{
Perform MBPCAOS
}
\examples{

data('antibiotic')
antb.uses <- antibiotic[,c('Atb.conso','Atb.Sys')]
health <- antibiotic[,c('Age','Loss')]
vet.practices <- antibiotic[,c(6:15)]
antibiotic <- data.frame(antb.uses,health,vet.practices)
# Defining blocks
blocs.name =  c("antibiotic.uses","Health.of.turkeys","Veterinary.practices")
blocs <- c(2,2,10)

# Level of scaling
level.scale <- rep(NA,ncol(antibiotic))
res.nature <- nature.variables(antibiotic)
level.scale [res.nature$p.numeric] <- "num"
level.scale [res.nature$p.quali] <- "nom"
#Warning; the ordinal nature of variables can not be detected automaticaly.
level.scale[c(1,14)] <- "ord"

res.MBPCAOS <- MBPCAOS(data = antibiotic,
                     level.scale = level.scale,
                      blocs = blocs,
                      blocs.name = blocs.name,
                     nb.comp = 3)
# Blocs graphs
plot.MBPCAOS(res.MBPCAOS,choice = 'blocs')




}
