% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.data.R
\name{mix.data}
\alias{mix.data}
\title{mix.data}
\usage{
mix.data(data)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}
}
\value{
\itemize{
  \item data.numeric : data.frame with numeric variables
  \item p.numeric : columns index of numeric variables
  \item nb.numeric : number of numeric variables
  \item data.quali : data.frame with numeric variables
  \item p.quali : columns index of numeric variables
  \item nb.quali :  number of numeric variables
}
}
\description{
Splits a mixed data matrix according to the nature of variables. Columns of class integer are considered quantitative. Character of factor are consider qualitative.
}
\examples{
data (antibiotic)
res.mix <- PCA.OS::mix.data(antibiotic)
res.mix$p.numeric
head(res.mix$data.numeric)
res.mix$p.quali
head(res.mix$data.quali)
nature <- rep(NA,ncol(antibiotic)) #Setting nature argument for PCAOS function
nature[res.mix$p.numeric] <- "num"
nature[res.mix$p.quali] <- "nom"
nature[c(1,15)] <- "ord"
nature
}
