% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.data.R
\name{nature.variables}
\alias{nature.variables}
\title{nature.variables}
\usage{
nature.variables(data, print.nature = TRUE)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{print.nature}{boolean (default = TRUE), if TRUE results are printed.}
}
\value{
\itemize{
  \item data.numeric : data.frame with numeric variables
  \item p.numeric : columns index of numeric variables
  \item nb.numeric : number of numeric variables
  \item data.quali : data.frame with numeric variables
  \item p.quali : columns index of numeric variables
  \item nb.quali :  number of numeric variables
}
}
\description{
Splits a data matrix according to the detected nature of variables. Columns of class integer are considered quantitative (num). Character of factor are consider qualitative (nom). Warning, ordinal variables can not be detected by this function, and must be manually indicated by the user.
}
\examples{
data (antibiotic)
res.mix <- PCA.OS::nature.variables(antibiotic)
res.mix$p.numeric
head(res.mix$data.numeric)
res.mix$p.quali
head(res.mix$data.quali)
nature <- rep(NA,ncol(antibiotic)) #Setting nature argument for PCAOS function
nature[res.mix$p.numeric] <- "num"
nature[res.mix$p.quali] <- "nom"
  nature[2] <- "ord"

}
