% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominal.quantification.R
\name{nominal.quantification}
\alias{nominal.quantification}
\title{nominal.quantification}
\usage{
nominal.quantification(var, t, rank.restriction)
}
\arguments{
\item{var}{raw nominal variables to be quantified}

\item{t}{components scores obtain trough PCAOS method}

\item{rank.restriction}{whether the quantification should be restricted or not}
}
\value{
\itemize{
  \item var.quant : optimally quantified variables
  \item w : loadings
  \item Yj : quantification of the categories
  \item Yjhat : rank one quantification of the categories
}
}
\description{
nominal.quantification
}
\details{
Do NOT use this function unless you are ME, a package developer, or a jedi user who really knows what is doing.
}
