% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.quantification.R
\name{numeric.quantification}
\alias{numeric.quantification}
\title{numeric.quantification}
\usage{
numeric.quantification(var, t, D)
}
\arguments{
\item{var}{raw numeric variables to be quantified}

\item{t}{components scores obtain trough PCAOS method}

\item{D}{degree assumed between variable and component}
}
\value{
\itemize{
  \item var.quant : optimally quantified variables
  \item w : loadings
  \item Yj : quantification of the categories
  \item Yjhat : rank one quantification of the categories
}
}
\description{
numeric.quantification
}
\details{
Do NOT use this function unless you are ME, a package developer, or a jedi user who really knows what is doing.
}
