% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.quantification.R
\name{ordinal.quantification}
\alias{ordinal.quantification}
\title{ordinal.quantification}
\usage{
ordinal.quantification(var, t)
}
\arguments{
\item{var}{raw ordinal variables to be quantified}

\item{t}{components scores obtain trough PCAOS method}
}
\value{
\itemize{
  \item var.quant : optimally quantified variables
  \item w : loadings
  \item Yj : quantification of the categories
  \item Yjhat : rank one quantification of the categories
}
}
\description{
ordinal.quantification
}
\details{
Do NOT use this function unless you are ME, a package developer, or a jedi user who really knows what is doing.
}
