% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAOS.R
\name{PCAOS}
\alias{PCAOS}
\title{Principal Component Analysis with Optimal Scaling}
\usage{
PCAOS(
  data,
  nature = rep("num", ncol(data)),
  nb.comp = 2,
  maxiter = 100,
  threshold = 1e-06,
  D = 1,
  rank.restriction = "one",
  supp.var = NULL,
  print.order = TRUE
)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num". The order of categories for an ordinal variable is indicated by it's level.}

\item{nb.comp}{number of components of the model (by default 2)}

\item{maxiter}{maximum number of iterations}

\item{threshold}{threshold for assessing convergence}

\item{D}{degree of the relation between quantified variables and components (only for numeric variables, default = 1)}

\item{rank.restriction}{restriction of the quantification matrix for nominal variable, possible values are :
\itemize{
  \item "one" (default) : apply a rank-one restriction (i.e single quantification of each variable)
  \item "no.restriction": apply no restriction (i.e multiple quantification of each variable)
}}

\item{supp.var}{a vector indicating the indexes of the supplementary variables}

\item{print.order}{boolean (TRUE by default), if TRUE ther order of the categories of ordinal variables are print}
}
\value{
\itemize{
  \item weigths : list of weights of the variables (loadings and weights are the same in PCA-like model)
  \item components : data.frame with individuals scores for each dimension
  \item quantified.data : Optimally quantified variables
  \item summary : summary of the number of variables according to their nature
  \item quantification.categories.nom : list of optimally quantified categories (nominal variables)
  \item quantification.categories.ord : list of optimally quantified categories (ordinal variables)
  \item inertia : percentage and cumulative percentage of variance of the quantified variables explained
  \item loss.tot : global loss for all variables
  \item stockiter : evolution of the criterion for each ieration
  \item data : orginal dataset
  \item nature : nature of scaling choosen for each variable
  \item quali.var.supp : supplementary variable

}
}
\description{
Perform PCAOS
}
\examples{
data (antibiotic)
nature <- rep(NA,ncol(antibiotic)) #Setting nature argument
nature[c(2,3,4)] <- "num"
nature[c(1,5,6,7,8,9,10,11,12,13,14,15)] <- "nom"
nature[c(1,15)] <- "ord"
nature

res.PCAOS <- PCA.OS::PCAOS(
 data = antibiotic,
 nature = nature,
 supp.var = 1
)

PCA.OS::plot.PCAOS(
 res.PCAOS = res.PCAOS,
 choice = "ind",
 coloring = antibiotic$Atb.conso)

}
\references{
Paries, Bougeard, Vigneau (2022), Multivariate analysis of Just-About-Right data with optimal scaling approach. Food Quality and Preference (submit)
}
\author{
\itemize{
  \item Martin PARIES (Maintainer: \email{martin.paries@oniris-nantes.fr})
  \item Evelyne Vigneau
  \item Stephanie Bougeard
}
}
