% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MBPCAOS.R
\name{plot.MBPCAOS}
\alias{plot.MBPCAOS}
\title{plot.MBPCAOS}
\usage{
\method{plot}{MBPCAOS}(
  x,
  choice,
  comp = c(1, 2),
  coloring.indiv = NULL,
  supp.var = FALSE,
  sub.var.quantif = NULL,
  sub.bloc = NULL,
  size.label = 3.5,
  size.legend = 12,
  label.cat = "var+cat",
  min.contribution = 0,
  ellipse = FALSE,
  level.conf = 0.95,
  ordinal.as.direction = FALSE,
  label.size.freq = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class MBPCAOS}

\item{choice}{the graph to plot possible values are "screeplot","quantif","indiv","cor","modalities","mixed","squared loadings". See Details.}

\item{comp}{a length 2 vector with the components to plot}

\item{coloring.indiv}{a vector of length N to color individuals. If NULL, no coloring is applied (individuals plot).}

\item{supp.var}{boolean (FALSE by default), if TRUE supplementary variables are added in factorial representation}

\item{sub.var.quantif}{a vector with variable of interest (quantification plots).}

\item{sub.bloc}{a scalar indicating the block to plot for variables graphs (i.e if sub.bloc == 1, variables of the first block are plotted) (all.var plot).}

\item{size.label}{size of label in graphs (all plots).}

\item{size.legend}{size of label in graphs (all plots).}

\item{label.cat}{if == 'var+cat', the name of the variable is included in the labels of the categories; if == 'cat', only the name of the categorie is plotted (name of categories should be unique) (qualitative and all.var plot)}

\item{min.contribution}{(all.var plot) Variables with a contribution (i.e loading) lower than this value will not be plotted in the 'all.var' graph (useful for dataset with a lot of variables)}

\item{ellipse}{boolean (FALSE by default), if TRUE, draw ellipses around categories of the qualitative variable considered as supplementary (individuals plot).}

\item{level.conf}{level of confidence ellipses (individuals plot).}

\item{ordinal.as.direction}{boolean (FALSE by default); if TRUE ordinal variables are represented as vectors, from the first categorie to the last one (qualitative and all.var plot).}

\item{label.size.freq}{boolean (FALSE by default); if TRUE size of categories are proportional to their citation frequencies (qualitative and all.var plot).}

\item{...}{further arguments passed to or from other methods, such as cex, cex.main, ...}
}
\value{
A ggplot object
}
\description{
Visualisation of results from MBPCAOS method. See details for available plots.
}
\details{
\itemize{
  \item  screeplot: Representation of the percentage of inertia restituates (Y), for each component (X).
  \item  quantif: Reprensetation of the quantification of variables trought Optimal Scaling, with original variables (X) and quantified variables (Y).Possibility to select one or more variables of interest with the argument "var.sub".
  \item  ind: factorial representation of individuals
  }

For numeric variables
  \itemize{
  \item  numeric: factorial representation of numeric variables (also called loading plot) Each numeric variable is represented by it's weight/loadings
  }
For qualitative (i.e nominal and ordinal) variables
\itemize{
  \item  qualitative: factorial representation of qualitatives variables trough the representation of it's categories. Coordinates of each category is calculted such as the single quantification of the category multiplied by the loading of the associated variable.
}
For mixed  variables
 \itemize{
  \item  all.var: factorial representation of all variables (weight for numeric variables, and categories for qualitative variables)
  \item  squared.loadings: plot of the squared loadings of all the variables.
}
For blocs
 \itemize{
  \item  blocks: contribution of each block to principal components. The contribution of one block is calculated as the squared sum of the loading of the variables in the block, divided by the block scaling of the block.
}

All graph are ggplot object
}
\examples{

#'data('antibiotic')
antb.uses <- antibiotic[,c('Atb.conso','Atb.Sys')]
health <- antibiotic[,c('Age','Loss')]
vet.practices <- antibiotic[,c(6:15)]
antibiotic <- data.frame(antb.uses,health,vet.practices)
# Defining blocks
blocks.name =  c("antibiotic.uses","Health.of.turkeys","Veterinary.practices")
blocks <- c(2,2,10)

# Level of scaling
level.scale <- rep(NA,ncol(antibiotic))
res.nature <- nature.variables(antibiotic)
level.scale [res.nature$p.numeric] <- "num"
level.scale [res.nature$p.quali] <- "nom"
#Warning; the ordinal nature of variables can not be detected automaticaly.
level.scale[c(1,14)] <- "ord"

# MBPCAOS
res.MBPCAOS <- MBPCAOS(data = antibiotic,
                     level.scale = level.scale,
                      blocks = blocks,
                      blocks.name = blocks.name,
                      nb.comp = 3)

# Blocks graphs
plot.MBPCAOS(x = res.MBPCAOS,choice = 'blocks')

}
