% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynom.data.R
\name{polynom.data}
\alias{polynom.data}
\title{polynom.data}
\usage{
polynom.data(var, D)
}
\arguments{
\item{var}{variable to pretreat}

\item{D}{Degree of the relation between quantified variables and components}
}
\value{
dummy coded variables
}
\description{
polynom.data
}
