% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibiotic.R
\docType{data}
\name{antibiotic}
\alias{antibiotic}
\title{Sample Data for Analysis}
\format{
A data frame with 122 row (agricultural explotation) and 15 columns (variables).
}
\usage{
antibiotic
}
\description{
The data used here refer epidemiologic measured.
}
\keyword{datasets}
