% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.component.R
\name{choice.component}
\alias{choice.component}
\title{choice.component}
\usage{
choice.component(
  data,
  nature = rep("num", ncol(data)),
  nb.comp.to.investigate = ncol(data),
  rank.restriction = "one"
)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{nb.comp.to.investigate}{Number of components to investigate (default=ncol(data))}

\item{rank.restriction}{rank restriction}
}
\value{
Data frame with global Loss value and percentage of the quantified variables inertia explained, for different model dimensionality (H). The last column provides the variation in percentage between to two successive size of dimensionality.
}
\description{
Helps the user to choose the appropriate number of PCAOS component to optimize
}
\examples{
data("antibiotic")

#Construction of the "nature" argument for this dataset
nature <- rep(NA,ncol(antibiotic)) #Setting nature argument
nature[c(2,3,4)] <- "num"
nature[c(1,5,6,7,8,9,10,11,12,13,14,15)] <- "nom"
nature[c(1,15)] <- "ord"

res.choice <- choice.component(antibiotic,nature)
res.choice

}
\author{
\itemize{
  \item Martin PARIES (Maintainer: \email{martin.paries@oniris-nantes.fr})
  \item Evelyne Vigneau
  \item Stephanie Bougeard
}
}
