% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBPCAOS.R
\name{MBPCAOS}
\alias{MBPCAOS}
\title{MultiBlock Principal Components Analysis with Optimal Scaling features}
\usage{
MBPCAOS(
  data,
  blocs,
  blocs.name = paste("Bloc", 1:length(blocs)),
  nature = rep("num", ncol(data)),
  block.scaling = "lambda1",
  nb.comp = 2,
  maxiter = 100,
  threshold = 1e-06,
  D = 1,
  rank.restriction = "one",
  supp.var = NULL
)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{blocs}{vector(length k) with number of variables in each bloc}

\item{blocs.name}{vector(length k) with names of each bloc}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{block.scaling}{scaling applied to each block. Possible value are : \itemize{
  \item "lambda1" (default) : each quantified block is divided by its the first singular value.
  \item "inertia": each quantified block is divided by its total inertia (sum of square).
  \item "null" : no scaling is applied
}}

\item{nb.comp}{number of components of the model (by default 2)}

\item{maxiter}{maximum number of iterations.}

\item{threshold}{the threshold for assessing convergence}

\item{D}{degree of the relation between quantified variables and components (only for numeric variables, default = 1)}

\item{rank.restriction}{number of components of the model (by default 2)}

\item{supp.var}{a vector indicating the indexes of the supplementary variables}
}
\value{
\itemize{
  \item weigths : list of weights of the variables (loadings and weights are the same in PCA like model)
  \item components : data.frame with individuals scores on each dimension.
  \item quantified.data : Optimally quantified variables
  \item quantified.data.list : Optimally quantified blocks (list form)
  \item block.components : components associated with each block
  \item summary : summary of number of variables and it's nature
  \item quantification.categories.nom : list of optimally quantified categories (nominal variables)
  \item quantification.categories.ord : list of optimally quantified categories (ordinal variables)
  \item inertia : percentage and cumulative percentage of variance of the quantified variables explained
  \item stockiter : evolution of criterion for each ieration
  \item data : orginal dataset
  \item nature : nature of scaling choosen for each variable
  \item blocs : number of variable in each block
  \item blocs.name : name of each block

}
}
\description{
Perform MBPCAOS
}
\examples{



}
