% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.OS.R
\docType{package}
\name{PCA.OS}
\alias{PCA.OS}
\title{PCA.OS}
\description{
Principal Component Analysis with Optimal Scaling features
}
\details{
\itemize{
  \item Package : PCA.OS
  \item Version : 0.1
  \item Date : 14/12/2021
  \item License :
}
}
\references{

}
\author{
\itemize{
  \item Martin PARIES (Maintainer: \email{martin.paries@oniris-nantes.fr})
  \item Evelyne Vigneau
  \item Stephanie Bougeard
}
}
