% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.OS.R
\docType{package}
\name{PCA.OS}
\alias{PCA.OS}
\title{PCA.OS}
\description{
The PCA.OS pacakge allows for single and multi-block exploratory analysis of variables of different nature. It allows both the dimension reduction and the quantification of datasets trough Optimal Scaling. Main functions are PCAOS for Principal Component Analysis with Optimal Scaling features and MBPCAOS for Multiblock version.
}
\details{
\itemize{
  \item Package : PCA.OS
  \item Version : 0.1
  \item Date : 14/12/2021
  \item License :
}
}
\references{
Paries, M., Bougeard, S., & Vigneau, E. (2022). Multivariate analysis of Just-About-Right data with optimal scaling approach. Food Quality and Preference, 102, 104681.
De Leeuw, J. (2013). History of nonlinear principal component analysis. eScholarship, University of California.
}
\author{
\itemize{
  \item Martin PARIES (Maintainer: \email{martin.paries@oniris-nantes.fr})
  \item Evelyne Vigneau
  \item Stephanie Bougeard
}
}
