% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAOS.R
\name{PCAOS}
\alias{PCAOS}
\title{Principal Component Analysis with Optimal Scaling}
\usage{
PCAOS(
  data,
  level.scale = rep("num", ncol(data)),
  nb.comp = 2,
  maxiter = 100,
  threshold = 1e-06,
  D = 1,
  rank.restriction = "one",
  supp.var = NULL,
  print.order = TRUE,
  init = "rdm"
)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{level.scale}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num". The order of categories for an ordinal variable is indicated by it's level.}

\item{nb.comp}{number of components of the model (by default 2)}

\item{maxiter}{maximum number of iterations}

\item{threshold}{threshold for assessing convergence}

\item{D}{degree of the relation between quantified variables and components (only for numeric variables, default = 1)}

\item{rank.restriction}{restriction of the quantification matrix for nominal variable, possible values are :
\itemize{
  \item "one" (default) : apply a rank-one restriction (i.e single quantification of each variable)
  \item "no.restriction": apply no restriction (i.e multiple quantification of each variable)
}}

\item{supp.var}{a vector indicating the indexes of the supplementary variables}

\item{print.order}{boolean (TRUE by default), if TRUE ther order of the categories of ordinal variables are print}

\item{init}{Intitialization strategy, possible values are :
\itemize{
  \item "rdm" (default) : random initialisation
  \item "svd": components are initialized with the singular value decomposition of the concatenated and pretreated variables (numeric variables are standardized and categorical variables are coded as pseudo disjunctive and weighted by the frequencies of the categories)
}}
}
\value{
\itemize{
  \item weigths : list of weights of the variables (loadings and weights are the same in PCA-like model)
  \item components : data.frame with individuals scores for each dimension
  \item quantified.data : optimally quantified variables
  \item summary : summary of the number of variables according to their nature
  \item quantification.categories.nom : list of optimally quantified categories (nominal variables)
  \item quantification.categories.ord : list of optimally quantified categories (ordinal variables)
  \item inertia : percentage and cumulative percentage of variance of the quantified variables explained
  \item loss.tot : global loss for all variables
  \item stockiter : evolution of the criterion for each ieration
  \item data : orginal dataset
  \item level.scale : nature of scaling choosen for each variable
  \item coord.supp.quali : coordinates of the supplementary qualitatives variables,
  \item coord.supp.num : coordinates of the supplementary numeric variables
  \item quali.var.supp : qualitative supplementary variable

}
}
\description{
Perform PCAOS
}
\examples{

data (antibiotic)
# Level of scaling of each variable
# Manually
level.scale <- rep(NA,ncol(antibiotic)) #Setting level.scale argument
level.scale[c(3,4)] <- "num"
level.scale[c(6:14)] <- "nom"
level.scale[c(1,15)] <- "ord"
# Or using nature.variables()
level.scale <- rep(NA,ncol(antibiotic))
res.nature <- nature.variables(antibiotic)
level.scale [res.nature$p.numeric] <- "num"
level.scale [res.nature$p.quali] <- "nom"
#Warning; the ordinal nature of variables can not be detected automaticaly.
level.scale[c(1,15)] <- "ord"

# PCAOS
res.PCAOS <- PCAOS(
data = antibiotic,
level.scale = level.scale,
rank.restriction = "one",
nb.comp = 2)

# Plot (individuals)
plot.PCAOS(
 res.PCAOS = res.PCAOS,
 choice = "ind",
 coloring.indiv = antibiotic$Atb.conso,
 size.legend = 12,
 size.label = 4
)

}
\references{
Paries, Bougeard, Vigneau (2022), Multivariate analysis of Just-About-Right data with optimal scaling approach. Food Quality and Preference (submit)
}
\author{
\itemize{
  \item Martin PARIES (Maintainer: \email{martin.paries@oniris-nantes.fr})
  \item Evelyne Vigneau
  \item Stephanie Bougeard
}
}
