% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAOS.R
\name{PCAOS}
\alias{PCAOS}
\title{Principal Components Analysis with Optimal Scaling features}
\usage{
PCAOS(
  data,
  nature = rep("num", ncol(data)),
  nb.comp = 2,
  maxiter = 100,
  threshold = 1e-06,
  D = 1,
  rank.restriction = "one",
  supp.var = NULL
)
}
\arguments{
\item{data}{a data frame with n rows (individuals) and p columns (numeric, nominal and/or ordinal variables)}

\item{nature}{vector(length p) giving the nature of each variable. Possible values: "nom", "ord", "num"}

\item{nb.comp}{number of components of the model (by default 2)}

\item{maxiter}{maximum number of iterations.}

\item{threshold}{the threshold for assessing convergence}

\item{D}{degree of the relation between quantified variables and components (only for numeric variables, default = 1)}

\item{rank.restriction}{restriction of the quantification matrix for nominal variable, possible values are :
\itemize{
  \item "one" (default) : apply a rank one restriction (i.e single quantification of each variable)
  \item "no.restriction": apply no restriction (i.e multiple quantification of each variable)
}}

\item{supp.var}{a vector indicating the indexes of the supplementary variables}
}
\value{
\itemize{
  \item weigths : list of weights of the variables (loadings and weights are the same in PCA like model)
  \item components : data.frame with individuals scores on each dimension.
  \item quantified.data : Optimally quantified variables
  \item summary : summary of number of variables and it's nature
  \item quantification.categories.nom : list of optimally quantified categories (nominal variables)
  \item quantification.categories.ord : list of optimally quantified categories (ordinal variables)
  \item inertia : percentage and cumulative percentage of variance of the quantified variables explained
  \item loss.tot : global loss for all variables
  \item stockiter : evolution of criterion for each ieration
  \item data : orginal dataset
  \item nature : nature of scaling choosen for each variable
  \item quali.var.supp : supplementary variable

}
}
\description{
Perform PCAOS
}
\examples{
data (antibiotic)
res.mix <- PCA.OS::mix.data(antibiotic)
nature <- rep(NA,ncol(antibiotic)) #Setting nature argument
nature[res.mix$p.numeric] <- "num"
nature[res.mix$p.quali] <- "nom"
nature[c(1,15)] <- "ord"
nature

res.PCAOS <- PCA.OS::PCAOS(
 data = antibiotic,
 nature = nature,
 supp.var = 1
)

PCA.OS::plot.PCAOS(res.PCAOS = res.PCAOS,choice = "ind",coloring = antibiotic$Atb.conso)

}
