% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MBPCAOS.R
\name{plot.MBPCAOS}
\alias{plot.MBPCAOS}
\title{plot.MBPCAOS}
\usage{
\method{plot}{MBPCAOS}(
  res.MBPCAOS,
  choice,
  comp = c(1, 2),
  coloring.ind = NULL,
  coloring.cat = "blocs",
  supp.var = FALSE,
  sub.var.quantif = NULL
)
}
\arguments{
\item{res.MBPCAOS}{an object of class MBPCAOS}

\item{choice}{the graph to plot possible values are "screeplot","quantif","indiv","cor","modalities","mixed","squared loadings". See Details.}

\item{comp}{a length 2 vector with the components to plot}

\item{coloring.ind}{A vector of length N to color individuals. If NULL, no coloring is applied.}

\item{coloring.cat}{Categories are colored according to their rescective blocs (default = "blocs"). If "variables", categories are colored according to their variables.}

\item{supp.var}{boolean (FALSE by default), if TRUE supplementary variables are added in factorial representation}

\item{sub.var.quantif}{a vector with variable of interest}
}
\value{
A ggplot object
}
\description{
Visualisation of results from MBPCAOS method. See details for available plot.
}
\details{
\itemize{
  \item  screeplot: Representation of the percentage of inertia restituates (Y), for each component (X).
  \item  quantif: Reprensetation of the quantification of variables trought Optimal Scaling, with original variables (X) and quantified variables (Y).Possibility to select one or more variables of interest with the argument "var.sub".
  \item  ind: factorial representation of individuals
  }

For numeric variables
  \itemize{
  \item  numeric: factorial representation of numeric variables (also called loading plot) Each numeric variable is represented by it's weight/loadings
  }
For qualitative (i.e nominal and ordinal) variables
\itemize{
  \item  qualitative: factorial representation of qualitatives variables trough the representation of it's categories. Coordinates of each category is calculted such as the single quantification of the category multiplied by the loading of the associated variable.
}
For mixed  variables
 \itemize{
  \item  mixed: factorial representation of mixed variables (weight for numeric variables, and categories for qualitative variables)
  \item  squared.loadings: plot of the squared loadings of all the variables.
}
For blocs
 \itemize{
  \item  blocs: factorial representation of blocs, trough the correlation of block component with principal components
}

All graph are ggplot object
}
